fix(tool): reprompt precedence for tool/hitl + route empty output via on_empty
P1: tool/hitl actions no longer inherit reprompt from workflow defaults.
reprompt is an LLM-recovery directive — re-running a deterministic UDF (or a
HITL step) on the same input can't change the outcome — so a defaults-level
reprompt must not bleed onto them. An explicit action-level reprompt is still
honoured. This removes the per-record 'reprompt is inert for tools' coercion
warning that fired once per record for every tool action.

P2: empty output is exempt from schema validation and routes via the on_empty
handler for all action types. An empty response has no fields to validate;
rejecting it in reject mode pre-empted the on_empty policy (skip/warn/error)
with a hard SchemaValidationError. _is_empty_output moves from the online
strategy down into the shared processing helpers (correct import direction)
and now gates schema validation.

Together these close the empty-tool-output crash: a deterministic tool that
legitimately returns nothing (e.g. everything filtered upstream) now reaches
its on_empty policy instead of crashing the run mid-pipeline.
