+++ agent_actions/processing/helpers.py
@@ -151,6 +151,20 @@ def _reject_schema_echo_items(response: Any, agent_name: str) -> Any:
     return result
 
 
+def _is_empty_output(response: Any) -> bool:
+    """Check if a tool/LLM response is effectively empty."""
+    if response is None:
+        return True
+    if isinstance(response, dict) and len(response) == 0:
+        return True
+    if isinstance(response, list):
+        if len(response) == 0:
+            return True
+        if all(isinstance(item, dict) and len(item) == 0 for item in response):
+            return True
+    return False
+
+
 def _validate_llm_output_schema(
     response: Any,
     agent_config: dict[str, Any],
@@ -183,6 +197,12 @@ def _validate_llm_output_schema(
                 agent_name,
                 mismatch_mode,
             )
+        return response
+
+    # Empty output is the on_empty handler's domain, not schema validation's:
+    # an empty response has no fields to check, so rejecting it here would
+    # pre-empt the on_empty policy (skip/warn/error) with a hard failure.
+    if _is_empty_output(response):
         return response
 
     mismatch_mode = _resolve_schema_mismatch_mode(agent_config)
